Update com_google_protobuf_javalite reference to use v3.11
The com_google_protobuf_javalite example in Readme.md uses an old java lite toolchain,
which is not compatible with the 3.11 runtime that is used by rules_proto. Update the
docs to have com_google_protobuf_javalite use the 3.11 toolchain so projects can build
with both Bazel 3.x and 2.x.